refact: change wallet terminology
