+++ package.json
@@ -58,7 +58,6 @@
     "eslint-config-airbnb-typescript": "^16.2.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "forc-bin": "workspace:*",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsdoc": "^37.9.7",
     "eslint-plugin-jsx-a11y": "^6.6.1",
@@ -67,6 +66,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "ethers": "^5.7.2",
+    "forc-bin": "workspace:*",
     "jest": "28.1.0",
     "markdownlint": "^0.23.1",
     "markdownlint-cli": "^0.27.1",
