+++ packages/wallet/package.json
@@ -24,16 +24,16 @@
   "dependencies": {
     "@ethersproject/bytes": "^5.7.0",
     "@fuel-ts/abi-coder": "workspace:*",
+    "@fuel-ts/address": "workspace:*",
     "@fuel-ts/constants": "workspace:*",
-    "@fuel-ts/math": "workspace:*",
     "@fuel-ts/hasher": "workspace:*",
     "@fuel-ts/hdwallet": "workspace:*",
     "@fuel-ts/interfaces": "workspace:*",
+    "@fuel-ts/math": "workspace:*",
     "@fuel-ts/mnemonic": "workspace:*",
     "@fuel-ts/providers": "workspace:*",
     "@fuel-ts/signer": "workspace:*",
-    "@fuel-ts/transactions": "workspace:*",
-    "@fuel-ts/address": "workspace:*"
+    "@fuel-ts/transactions": "workspace:*"
   },
   "devDependencies": {
     "@fuel-ts/keystore": "workspace:*",
