+++ packages/wallet/src/constants.ts
@@ -1,0 +1,2 @@
+// TODO: import using .env file
+export const FUEL_NETWORK_URL = 'http://127.0.0.1:4000/graphql';
